Misc changes, checked in some state files from Untangle, unittest for GymNode.

+++ LogLib/unittest/gasnodetest.cpp
@@ -1,6 +1,8 @@
 #include "GasNode.h"
 #include "Node.h"
 #include <string>
+#include <stdio.h>
+#include <time.h>
 using namespace std;
 
 int main()
@@ -15,7 +17,7 @@ int main()
   str = "1234 1234 1.234 1.234";
   printf("\t'%s' (expect yes): %s\n", str.c_str(), n->setString(str) ? "yes" : "no");
   printf("date = %d, miles = %d, price = %f gallons = %f;; total = %f\n",
-            n->getDate(),
+            (int)n->getDate(),
             n->getMiles(),
             n->pricePerGallon(),
             n->getGallons(),
@@ -24,7 +26,7 @@ int main()
   str = "1 1 1 1";
   printf("\t'%s' (expect yes): %s\n", str.c_str(), n->setString(str) ? "yes" : "no");
   printf("date = %d, miles = %d, price = %f gallons = %f;; total = %f\n",
-            n->getDate(),
+            (int)n->getDate(),
             n->getMiles(),
             n->pricePerGallon(),
             n->getGallons(),
@@ -33,7 +35,7 @@ int main()
   str = "1234567 454346346  1.25 4";
   printf("\t'%s' (expect yes): %s\n", str.c_str(), n->setString(str) ? "yes" : "no");
   printf("date = %d, miles = %d, price = %f gallons = %f;; total = %f\n",
-            n->getDate(),
+            (int)n->getDate(),
             n->getMiles(),
             n->pricePerGallon(),
             n->getGallons(),
@@ -42,7 +44,7 @@ int main()
   str = "-5 -5 -5 -5";
   printf("\t'%s' (expect yes): %s\n", str.c_str(), n->setString(str) ? "yes" : "no");
   printf("date = %d, miles = %d, price = %f gallons = %f;; total = %f\n",
-            n->getDate(),
+            (int)n->getDate(),
             n->getMiles(),
             n->pricePerGallon(),
             n->getGallons(),
